Refactor driver module names
