+++ squirrel/driver/__init__.py
@@ -1,15 +1,15 @@
-from squirrel.driver.csv_driver import CsvDriver
-from squirrel.driver.data_frame_file_driver import DataFrameFileDriver
+from squirrel.driver.csv import CsvDriver
+from squirrel.driver.data_frame_file import DataFrameFileDriver
 from squirrel.driver.driver import DataFrameDriver, Driver, IterDriver, MapDriver
-from squirrel.driver.file_driver import FileDriver
-from squirrel.driver.feather_driver import FeatherDriver
-from squirrel.driver.json_driver import JsonDriver
+from squirrel.driver.file import FileDriver
+from squirrel.driver.feather import FeatherDriver
+from squirrel.driver.json import JsonDriver
 from squirrel.driver.jsonl import JsonlDriver
 from squirrel.driver.msgpack import MessagepackDriver
-from squirrel.driver.parquet_driver import ParquetDriver
+from squirrel.driver.parquet import ParquetDriver
 from squirrel.driver.source_combiner import SourceCombiner
-from squirrel.driver.store_driver import StoreDriver
-from squirrel.driver.xls_driver import XlsDriver
+from squirrel.driver.store import StoreDriver
+from squirrel.driver.xls import XlsDriver
 from squirrel.driver.zarr import ZarrDriver
 
 __all__ = [
